Rename image name and fix tests
