+++ vllm-cuda-gpu/rock-ci-metadata.yaml
